@@ -4,8 +4,8 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.36.9
+// 	protoc        v6.32.0
 // source: compact_formats.proto
 
 package walletrpc
@@ -15,6 +15,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -26,21 +27,18 @@ const (
 
 // ChainMetadata represents information about the state of the chain as of a given block.
 type ChainMetadata struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	SaplingCommitmentTreeSize uint32 `protobuf:"varint,1,opt,name=saplingCommitmentTreeSize,proto3" json:"saplingCommitmentTreeSize,omitempty"` // the size of the Sapling note commitment tree as of the end of this block
-	OrchardCommitmentTreeSize uint32 `protobuf:"varint,2,opt,name=orchardCommitmentTreeSize,proto3" json:"orchardCommitmentTreeSize,omitempty"` // the size of the Orchard note commitment tree as of the end of this block
+	state                     protoimpl.MessageState `protogen:"open.v1"`
+	SaplingCommitmentTreeSize uint32                 `protobuf:"varint,1,opt,name=saplingCommitmentTreeSize,proto3" json:"saplingCommitmentTreeSize,omitempty"` // the size of the Sapling note commitment tree as of the end of this block
+	OrchardCommitmentTreeSize uint32                 `protobuf:"varint,2,opt,name=orchardCommitmentTreeSize,proto3" json:"orchardCommitmentTreeSize,omitempty"` // the size of the Orchard note commitment tree as of the end of this block
+	unknownFields             protoimpl.UnknownFields
+	sizeCache                 protoimpl.SizeCache
 }
 
 func (x *ChainMetadata) Reset() {
 	*x = ChainMetadata{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_compact_formats_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_compact_formats_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ChainMetadata) String() string {
@@ -51,7 +49,7 @@ func (*ChainMetadata) ProtoMessage() {}
 
 func (x *ChainMetadata) ProtoReflect() protoreflect.Message {
 	mi := &file_compact_formats_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -85,27 +83,24 @@ func (x *ChainMetadata) GetOrchardCommitmentTreeSize() uint32 {
 //  2. Detect a spend of your shielded Sapling notes
 //  3. Update your witnesses to generate new Sapling spend proofs.
 type CompactBlock struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	ProtoVersion  uint32                 `protobuf:"varint,1,opt,name=protoVersion,proto3" json:"protoVersion,omitempty"`  // the version of this wire format, for storage
+	Height        uint64                 `protobuf:"varint,2,opt,name=height,proto3" json:"height,omitempty"`              // the height of this block
+	Hash          []byte                 `protobuf:"bytes,3,opt,name=hash,proto3" json:"hash,omitempty"`                   // the ID (hash) of this block, same as in block explorers
+	PrevHash      []byte                 `protobuf:"bytes,4,opt,name=prevHash,proto3" json:"prevHash,omitempty"`           // the ID (hash) of this block's predecessor
+	Time          uint32                 `protobuf:"varint,5,opt,name=time,proto3" json:"time,omitempty"`                  // Unix epoch time when the block was mined
+	Header        []byte                 `protobuf:"bytes,6,opt,name=header,proto3" json:"header,omitempty"`               // (hash, prevHash, and time) OR (full header)
+	Vtx           []*CompactTx           `protobuf:"bytes,7,rep,name=vtx,proto3" json:"vtx,omitempty"`                     // zero or more compact transactions from this block
+	ChainMetadata *ChainMetadata         `protobuf:"bytes,8,opt,name=chainMetadata,proto3" json:"chainMetadata,omitempty"` // information about the state of the chain as of this block
 	unknownFields protoimpl.UnknownFields
-
-	ProtoVersion  uint32         `protobuf:"varint,1,opt,name=protoVersion,proto3" json:"protoVersion,omitempty"`  // the version of this wire format, for storage
-	Height        uint64         `protobuf:"varint,2,opt,name=height,proto3" json:"height,omitempty"`              // the height of this block
-	Hash          []byte         `protobuf:"bytes,3,opt,name=hash,proto3" json:"hash,omitempty"`                   // the ID (hash) of this block, same as in block explorers
-	PrevHash      []byte         `protobuf:"bytes,4,opt,name=prevHash,proto3" json:"prevHash,omitempty"`           // the ID (hash) of this block's predecessor
-	Time          uint32         `protobuf:"varint,5,opt,name=time,proto3" json:"time,omitempty"`                  // Unix epoch time when the block was mined
-	Header        []byte         `protobuf:"bytes,6,opt,name=header,proto3" json:"header,omitempty"`               // (hash, prevHash, and time) OR (full header)
-	Vtx           []*CompactTx   `protobuf:"bytes,7,rep,name=vtx,proto3" json:"vtx,omitempty"`                     // zero or more compact transactions from this block
-	ChainMetadata *ChainMetadata `protobuf:"bytes,8,opt,name=chainMetadata,proto3" json:"chainMetadata,omitempty"` // information about the state of the chain as of this block
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CompactBlock) Reset() {
 	*x = CompactBlock{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_compact_formats_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_compact_formats_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CompactBlock) String() string {
@@ -116,7 +111,7 @@ func (*CompactBlock) ProtoMessage() {}
 
 func (x *CompactBlock) ProtoReflect() protoreflect.Message {
 	mi := &file_compact_formats_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -191,10 +186,7 @@ func (x *CompactBlock) GetChainMetadata() *ChainMetadata {
 // is relevant to it (either pays to it or spends from it) via shielded elements
 // only. This message will not encode a transparent-to-transparent transaction.
 type CompactTx struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// Index and hash will allow the receiver to call out to chain
 	// explorers or other data structures to retrieve more information
 	// about this transaction.
@@ -206,19 +198,19 @@ type CompactTx struct {
 	// If there are no transparent inputs, the fee will be calculable as:
 	//
 	//	valueBalanceSapling + valueBalanceOrchard + sum(vPubNew) - sum(vPubOld) - sum(tOut)
-	Fee     uint32                  `protobuf:"varint,3,opt,name=fee,proto3" json:"fee,omitempty"`
-	Spends  []*CompactSaplingSpend  `protobuf:"bytes,4,rep,name=spends,proto3" json:"spends,omitempty"`
-	Outputs []*CompactSaplingOutput `protobuf:"bytes,5,rep,name=outputs,proto3" json:"outputs,omitempty"`
-	Actions []*CompactOrchardAction `protobuf:"bytes,6,rep,name=actions,proto3" json:"actions,omitempty"`
+	Fee           uint32                  `protobuf:"varint,3,opt,name=fee,proto3" json:"fee,omitempty"`
+	Spends        []*CompactSaplingSpend  `protobuf:"bytes,4,rep,name=spends,proto3" json:"spends,omitempty"`
+	Outputs       []*CompactSaplingOutput `protobuf:"bytes,5,rep,name=outputs,proto3" json:"outputs,omitempty"`
+	Actions       []*CompactOrchardAction `protobuf:"bytes,6,rep,name=actions,proto3" json:"actions,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CompactTx) Reset() {
 	*x = CompactTx{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_compact_formats_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_compact_formats_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CompactTx) String() string {
@@ -229,7 +221,7 @@ func (*CompactTx) ProtoMessage() {}
 
 func (x *CompactTx) ProtoReflect() protoreflect.Message {
 	mi := &file_compact_formats_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -289,20 +281,17 @@ func (x *CompactTx) GetActions() []*CompactOrchardAction {
 // CompactSaplingSpend is a Sapling Spend Description as described in 7.3 of the Zcash
 // protocol specification.
 type CompactSaplingSpend struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Nf            []byte                 `protobuf:"bytes,1,opt,name=nf,proto3" json:"nf,omitempty"` // nullifier (see the Zcash protocol specification)
 	unknownFields protoimpl.UnknownFields
-
-	Nf []byte `protobuf:"bytes,1,opt,name=nf,proto3" json:"nf,omitempty"` // nullifier (see the Zcash protocol specification)
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CompactSaplingSpend) Reset() {
 	*x = CompactSaplingSpend{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_compact_formats_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_compact_formats_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CompactSaplingSpend) String() string {
@@ -313,7 +302,7 @@ func (*CompactSaplingSpend) ProtoMessage() {}
 
 func (x *CompactSaplingSpend) ProtoReflect() protoreflect.Message {
 	mi := &file_compact_formats_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -341,22 +330,19 @@ func (x *CompactSaplingSpend) GetNf() []byte {
 // https://zips.z.cash/protocol/protocol.pdf#outputencodingandconsensus
 // Total size is 116 bytes.
 type CompactSaplingOutput struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Cmu           []byte                 `protobuf:"bytes,1,opt,name=cmu,proto3" json:"cmu,omitempty"`                   // note commitment u-coordinate
+	EphemeralKey  []byte                 `protobuf:"bytes,2,opt,name=ephemeralKey,proto3" json:"ephemeralKey,omitempty"` // ephemeral public key
+	Ciphertext    []byte                 `protobuf:"bytes,3,opt,name=ciphertext,proto3" json:"ciphertext,omitempty"`     // first 52 bytes of ciphertext
 	unknownFields protoimpl.UnknownFields
-
-	Cmu          []byte `protobuf:"bytes,1,opt,name=cmu,proto3" json:"cmu,omitempty"`                   // note commitment u-coordinate
-	EphemeralKey []byte `protobuf:"bytes,2,opt,name=ephemeralKey,proto3" json:"ephemeralKey,omitempty"` // ephemeral public key
-	Ciphertext   []byte `protobuf:"bytes,3,opt,name=ciphertext,proto3" json:"ciphertext,omitempty"`     // first 52 bytes of ciphertext
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CompactSaplingOutput) Reset() {
 	*x = CompactSaplingOutput{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_compact_formats_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_compact_formats_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CompactSaplingOutput) String() string {
@@ -367,7 +353,7 @@ func (*CompactSaplingOutput) ProtoMessage() {}
 
 func (x *CompactSaplingOutput) ProtoReflect() protoreflect.Message {
 	mi := &file_compact_formats_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -406,23 +392,20 @@ func (x *CompactSaplingOutput) GetCiphertext() []byte {
 // https://github.com/zcash/zips/blob/main/zip-0225.rst#orchard-action-description-orchardaction
 // (but not all fields are needed)
 type CompactOrchardAction struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Nullifier     []byte                 `protobuf:"bytes,1,opt,name=nullifier,proto3" json:"nullifier,omitempty"`       // [32] The nullifier of the input note
+	Cmx           []byte                 `protobuf:"bytes,2,opt,name=cmx,proto3" json:"cmx,omitempty"`                   // [32] The x-coordinate of the note commitment for the output note
+	EphemeralKey  []byte                 `protobuf:"bytes,3,opt,name=ephemeralKey,proto3" json:"ephemeralKey,omitempty"` // [32] An encoding of an ephemeral Pallas public key
+	Ciphertext    []byte                 `protobuf:"bytes,4,opt,name=ciphertext,proto3" json:"ciphertext,omitempty"`     // [52] The first 52 bytes of the encCiphertext field
 	unknownFields protoimpl.UnknownFields
-
-	Nullifier    []byte `protobuf:"bytes,1,opt,name=nullifier,proto3" json:"nullifier,omitempty"`       // [32] The nullifier of the input note
-	Cmx          []byte `protobuf:"bytes,2,opt,name=cmx,proto3" json:"cmx,omitempty"`                   // [32] The x-coordinate of the note commitment for the output note
-	EphemeralKey []byte `protobuf:"bytes,3,opt,name=ephemeralKey,proto3" json:"ephemeralKey,omitempty"` // [32] An encoding of an ephemeral Pallas public key
-	Ciphertext   []byte `protobuf:"bytes,4,opt,name=ciphertext,proto3" json:"ciphertext,omitempty"`     // [52] The first 52 bytes of the encCiphertext field
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CompactOrchardAction) Reset() {
 	*x = CompactOrchardAction{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_compact_formats_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_compact_formats_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CompactOrchardAction) String() string {
@@ -433,7 +416,7 @@ func (*CompactOrchardAction) ProtoMessage() {}
 
 func (x *CompactOrchardAction) ProtoReflect() protoreflect.Message {
 	mi := &file_compact_formats_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -478,92 +461,58 @@ func (x *CompactOrchardAction) GetCiphertext() []byte {
 
 var File_compact_formats_proto protoreflect.FileDescriptor
 
-var file_compact_formats_proto_rawDesc = []byte{
-	0x0a, 0x15, 0x63, 0x6f, 0x6d, 0x70, 0x61, 0x63, 0x74, 0x5f, 0x66, 0x6f, 0x72, 0x6d, 0x61, 0x74,
-	0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x15, 0x63, 0x61, 0x73, 0x68, 0x2e, 0x7a, 0x2e,
-	0x77, 0x61, 0x6c, 0x6c, 0x65, 0x74, 0x2e, 0x73, 0x64, 0x6b, 0x2e, 0x72, 0x70, 0x63, 0x22, 0x8b,
-	0x01, 0x0a, 0x0d, 0x43, 0x68, 0x61, 0x69, 0x6e, 0x4d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61,
-	0x12, 0x3c, 0x0a, 0x19, 0x73, 0x61, 0x70, 0x6c, 0x69, 0x6e, 0x67, 0x43, 0x6f, 0x6d, 0x6d, 0x69,
-	0x74, 0x6d, 0x65, 0x6e, 0x74, 0x54, 0x72, 0x65, 0x65, 0x53, 0x69, 0x7a, 0x65, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x0d, 0x52, 0x19, 0x73, 0x61, 0x70, 0x6c, 0x69, 0x6e, 0x67, 0x43, 0x6f, 0x6d, 0x6d,
-	0x69, 0x74, 0x6d, 0x65, 0x6e, 0x74, 0x54, 0x72, 0x65, 0x65, 0x53, 0x69, 0x7a, 0x65, 0x12, 0x3c,
-	0x0a, 0x19, 0x6f, 0x72, 0x63, 0x68, 0x61, 0x72, 0x64, 0x43, 0x6f, 0x6d, 0x6d, 0x69, 0x74, 0x6d,
-	0x65, 0x6e, 0x74, 0x54, 0x72, 0x65, 0x65, 0x53, 0x69, 0x7a, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28,
-	0x0d, 0x52, 0x19, 0x6f, 0x72, 0x63, 0x68, 0x61, 0x72, 0x64, 0x43, 0x6f, 0x6d, 0x6d, 0x69, 0x74,
-	0x6d, 0x65, 0x6e, 0x74, 0x54, 0x72, 0x65, 0x65, 0x53, 0x69, 0x7a, 0x65, 0x22, 0xa6, 0x02, 0x0a,
-	0x0c, 0x43, 0x6f, 0x6d, 0x70, 0x61, 0x63, 0x74, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x12, 0x22, 0x0a,
-	0x0c, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x0d, 0x52, 0x0c, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f,
-	0x6e, 0x12, 0x16, 0x0a, 0x06, 0x68, 0x65, 0x69, 0x67, 0x68, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28,
-	0x04, 0x52, 0x06, 0x68, 0x65, 0x69, 0x67, 0x68, 0x74, 0x12, 0x12, 0x0a, 0x04, 0x68, 0x61, 0x73,
-	0x68, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x04, 0x68, 0x61, 0x73, 0x68, 0x12, 0x1a, 0x0a,
-	0x08, 0x70, 0x72, 0x65, 0x76, 0x48, 0x61, 0x73, 0x68, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0c, 0x52,
-	0x08, 0x70, 0x72, 0x65, 0x76, 0x48, 0x61, 0x73, 0x68, 0x12, 0x12, 0x0a, 0x04, 0x74, 0x69, 0x6d,
-	0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x04, 0x74, 0x69, 0x6d, 0x65, 0x12, 0x16, 0x0a,
-	0x06, 0x68, 0x65, 0x61, 0x64, 0x65, 0x72, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x06, 0x68,
-	0x65, 0x61, 0x64, 0x65, 0x72, 0x12, 0x32, 0x0a, 0x03, 0x76, 0x74, 0x78, 0x18, 0x07, 0x20, 0x03,
-	0x28, 0x0b, 0x32, 0x20, 0x2e, 0x63, 0x61, 0x73, 0x68, 0x2e, 0x7a, 0x2e, 0x77, 0x61, 0x6c, 0x6c,
-	0x65, 0x74, 0x2e, 0x73, 0x64, 0x6b, 0x2e, 0x72, 0x70, 0x63, 0x2e, 0x43, 0x6f, 0x6d, 0x70, 0x61,
-	0x63, 0x74, 0x54, 0x78, 0x52, 0x03, 0x76, 0x74, 0x78, 0x12, 0x4a, 0x0a, 0x0d, 0x63, 0x68, 0x61,
-	0x69, 0x6e, 0x4d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x18, 0x08, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x24, 0x2e, 0x63, 0x61, 0x73, 0x68, 0x2e, 0x7a, 0x2e, 0x77, 0x61, 0x6c, 0x6c, 0x65, 0x74,
-	0x2e, 0x73, 0x64, 0x6b, 0x2e, 0x72, 0x70, 0x63, 0x2e, 0x43, 0x68, 0x61, 0x69, 0x6e, 0x4d, 0x65,
-	0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0x52, 0x0d, 0x63, 0x68, 0x61, 0x69, 0x6e, 0x4d, 0x65, 0x74,
-	0x61, 0x64, 0x61, 0x74, 0x61, 0x22, 0x99, 0x02, 0x0a, 0x09, 0x43, 0x6f, 0x6d, 0x70, 0x61, 0x63,
-	0x74, 0x54, 0x78, 0x12, 0x14, 0x0a, 0x05, 0x69, 0x6e, 0x64, 0x65, 0x78, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x04, 0x52, 0x05, 0x69, 0x6e, 0x64, 0x65, 0x78, 0x12, 0x12, 0x0a, 0x04, 0x68, 0x61, 0x73,
-	0x68, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x04, 0x68, 0x61, 0x73, 0x68, 0x12, 0x10, 0x0a,
-	0x03, 0x66, 0x65, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x03, 0x66, 0x65, 0x65, 0x12,
-	0x42, 0x0a, 0x06, 0x73, 0x70, 0x65, 0x6e, 0x64, 0x73, 0x18, 0x04, 0x20, 0x03, 0x28, 0x0b, 0x32,
-	0x2a, 0x2e, 0x63, 0x61, 0x73, 0x68, 0x2e, 0x7a, 0x2e, 0x77, 0x61, 0x6c, 0x6c, 0x65, 0x74, 0x2e,
-	0x73, 0x64, 0x6b, 0x2e, 0x72, 0x70, 0x63, 0x2e, 0x43, 0x6f, 0x6d, 0x70, 0x61, 0x63, 0x74, 0x53,
-	0x61, 0x70, 0x6c, 0x69, 0x6e, 0x67, 0x53, 0x70, 0x65, 0x6e, 0x64, 0x52, 0x06, 0x73, 0x70, 0x65,
-	0x6e, 0x64, 0x73, 0x12, 0x45, 0x0a, 0x07, 0x6f, 0x75, 0x74, 0x70, 0x75, 0x74, 0x73, 0x18, 0x05,
-	0x20, 0x03, 0x28, 0x0b, 0x32, 0x2b, 0x2e, 0x63, 0x61, 0x73, 0x68, 0x2e, 0x7a, 0x2e, 0x77, 0x61,
-	0x6c, 0x6c, 0x65, 0x74, 0x2e, 0x73, 0x64, 0x6b, 0x2e, 0x72, 0x70, 0x63, 0x2e, 0x43, 0x6f, 0x6d,
-	0x70, 0x61, 0x63, 0x74, 0x53, 0x61, 0x70, 0x6c, 0x69, 0x6e, 0x67, 0x4f, 0x75, 0x74, 0x70, 0x75,
-	0x74, 0x52, 0x07, 0x6f, 0x75, 0x74, 0x70, 0x75, 0x74, 0x73, 0x12, 0x45, 0x0a, 0x07, 0x61, 0x63,
-	0x74, 0x69, 0x6f, 0x6e, 0x73, 0x18, 0x06, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x2b, 0x2e, 0x63, 0x61,
-	0x73, 0x68, 0x2e, 0x7a, 0x2e, 0x77, 0x61, 0x6c, 0x6c, 0x65, 0x74, 0x2e, 0x73, 0x64, 0x6b, 0x2e,
-	0x72, 0x70, 0x63, 0x2e, 0x43, 0x6f, 0x6d, 0x70, 0x61, 0x63, 0x74, 0x4f, 0x72, 0x63, 0x68, 0x61,
-	0x72, 0x64, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x07, 0x61, 0x63, 0x74, 0x69, 0x6f, 0x6e,
-	0x73, 0x22, 0x25, 0x0a, 0x13, 0x43, 0x6f, 0x6d, 0x70, 0x61, 0x63, 0x74, 0x53, 0x61, 0x70, 0x6c,
-	0x69, 0x6e, 0x67, 0x53, 0x70, 0x65, 0x6e, 0x64, 0x12, 0x0e, 0x0a, 0x02, 0x6e, 0x66, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x0c, 0x52, 0x02, 0x6e, 0x66, 0x22, 0x6c, 0x0a, 0x14, 0x43, 0x6f, 0x6d, 0x70,
-	0x61, 0x63, 0x74, 0x53, 0x61, 0x70, 0x6c, 0x69, 0x6e, 0x67, 0x4f, 0x75, 0x74, 0x70, 0x75, 0x74,
-	0x12, 0x10, 0x0a, 0x03, 0x63, 0x6d, 0x75, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x03, 0x63,
-	0x6d, 0x75, 0x12, 0x22, 0x0a, 0x0c, 0x65, 0x70, 0x68, 0x65, 0x6d, 0x65, 0x72, 0x61, 0x6c, 0x4b,
-	0x65, 0x79, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x0c, 0x65, 0x70, 0x68, 0x65, 0x6d, 0x65,
-	0x72, 0x61, 0x6c, 0x4b, 0x65, 0x79, 0x12, 0x1e, 0x0a, 0x0a, 0x63, 0x69, 0x70, 0x68, 0x65, 0x72,
-	0x74, 0x65, 0x78, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x0a, 0x63, 0x69, 0x70, 0x68,
-	0x65, 0x72, 0x74, 0x65, 0x78, 0x74, 0x22, 0x8a, 0x01, 0x0a, 0x14, 0x43, 0x6f, 0x6d, 0x70, 0x61,
-	0x63, 0x74, 0x4f, 0x72, 0x63, 0x68, 0x61, 0x72, 0x64, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x12,
-	0x1c, 0x0a, 0x09, 0x6e, 0x75, 0x6c, 0x6c, 0x69, 0x66, 0x69, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x0c, 0x52, 0x09, 0x6e, 0x75, 0x6c, 0x6c, 0x69, 0x66, 0x69, 0x65, 0x72, 0x12, 0x10, 0x0a,
-	0x03, 0x63, 0x6d, 0x78, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x03, 0x63, 0x6d, 0x78, 0x12,
-	0x22, 0x0a, 0x0c, 0x65, 0x70, 0x68, 0x65, 0x6d, 0x65, 0x72, 0x61, 0x6c, 0x4b, 0x65, 0x79, 0x18,
-	0x03, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x0c, 0x65, 0x70, 0x68, 0x65, 0x6d, 0x65, 0x72, 0x61, 0x6c,
-	0x4b, 0x65, 0x79, 0x12, 0x1e, 0x0a, 0x0a, 0x63, 0x69, 0x70, 0x68, 0x65, 0x72, 0x74, 0x65, 0x78,
-	0x74, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x0a, 0x63, 0x69, 0x70, 0x68, 0x65, 0x72, 0x74,
-	0x65, 0x78, 0x74, 0x42, 0x1b, 0x5a, 0x16, 0x6c, 0x69, 0x67, 0x68, 0x74, 0x77, 0x61, 0x6c, 0x6c,
-	0x65, 0x74, 0x64, 0x2f, 0x77, 0x61, 0x6c, 0x6c, 0x65, 0x74, 0x72, 0x70, 0x63, 0xba, 0x02, 0x00,
-	0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_compact_formats_proto_rawDesc = "" +
+	"\n" +
+	"\x15compact_formats.proto\x12\x15cash.z.wallet.sdk.rpc\"\x8b\x01\n" +
+	"\rChainMetadata\x12<\n" +
+	"\x19saplingCommitmentTreeSize\x18\x01 \x01(\rR\x19saplingCommitmentTreeSize\x12<\n" +
+	"\x19orchardCommitmentTreeSize\x18\x02 \x01(\rR\x19orchardCommitmentTreeSize\"\xa6\x02\n" +
+	"\fCompactBlock\x12\"\n" +
+	"\fprotoVersion\x18\x01 \x01(\rR\fprotoVersion\x12\x16\n" +
+	"\x06height\x18\x02 \x01(\x04R\x06height\x12\x12\n" +
+	"\x04hash\x18\x03 \x01(\fR\x04hash\x12\x1a\n" +
+	"\bprevHash\x18\x04 \x01(\fR\bprevHash\x12\x12\n" +
+	"\x04time\x18\x05 \x01(\rR\x04time\x12\x16\n" +
+	"\x06header\x18\x06 \x01(\fR\x06header\x122\n" +
+	"\x03vtx\x18\a \x03(\v2 .cash.z.wallet.sdk.rpc.CompactTxR\x03vtx\x12J\n" +
+	"\rchainMetadata\x18\b \x01(\v2$.cash.z.wallet.sdk.rpc.ChainMetadataR\rchainMetadata\"\x99\x02\n" +
+	"\tCompactTx\x12\x14\n" +
+	"\x05index\x18\x01 \x01(\x04R\x05index\x12\x12\n" +
+	"\x04hash\x18\x02 \x01(\fR\x04hash\x12\x10\n" +
+	"\x03fee\x18\x03 \x01(\rR\x03fee\x12B\n" +
+	"\x06spends\x18\x04 \x03(\v2*.cash.z.wallet.sdk.rpc.CompactSaplingSpendR\x06spends\x12E\n" +
+	"\aoutputs\x18\x05 \x03(\v2+.cash.z.wallet.sdk.rpc.CompactSaplingOutputR\aoutputs\x12E\n" +
+	"\aactions\x18\x06 \x03(\v2+.cash.z.wallet.sdk.rpc.CompactOrchardActionR\aactions\"%\n" +
+	"\x13CompactSaplingSpend\x12\x0e\n" +
+	"\x02nf\x18\x01 \x01(\fR\x02nf\"l\n" +
+	"\x14CompactSaplingOutput\x12\x10\n" +
+	"\x03cmu\x18\x01 \x01(\fR\x03cmu\x12\"\n" +
+	"\fephemeralKey\x18\x02 \x01(\fR\fephemeralKey\x12\x1e\n" +
+	"\n" +
+	"ciphertext\x18\x03 \x01(\fR\n" +
+	"ciphertext\"\x8a\x01\n" +
+	"\x14CompactOrchardAction\x12\x1c\n" +
+	"\tnullifier\x18\x01 \x01(\fR\tnullifier\x12\x10\n" +
+	"\x03cmx\x18\x02 \x01(\fR\x03cmx\x12\"\n" +
+	"\fephemeralKey\x18\x03 \x01(\fR\fephemeralKey\x12\x1e\n" +
+	"\n" +
+	"ciphertext\x18\x04 \x01(\fR\n" +
+	"ciphertextB\x1bZ\x16lightwalletd/walletrpc\xba\x02\x00b\x06proto3"
 
 var (
 	file_compact_formats_proto_rawDescOnce sync.Once
-	file_compact_formats_proto_rawDescData = file_compact_formats_proto_rawDesc
+	file_compact_formats_proto_rawDescData []byte
 )
 
 func file_compact_formats_proto_rawDescGZIP() []byte {
 	file_compact_formats_proto_rawDescOnce.Do(func() {
-		file_compact_formats_proto_rawDescData = protoimpl.X.CompressGZIP(file_compact_formats_proto_rawDescData)
+		file_compact_formats_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_compact_formats_proto_rawDesc), len(file_compact_formats_proto_rawDesc)))
 	})
 	return file_compact_formats_proto_rawDescData
 }
 
 var file_compact_formats_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_compact_formats_proto_goTypes = []interface{}{
+var file_compact_formats_proto_goTypes = []any{
 	(*ChainMetadata)(nil),        // 0: cash.z.wallet.sdk.rpc.ChainMetadata
 	(*CompactBlock)(nil),         // 1: cash.z.wallet.sdk.rpc.CompactBlock
 	(*CompactTx)(nil),            // 2: cash.z.wallet.sdk.rpc.CompactTx
@@ -589,85 +538,11 @@ func file_compact_formats_proto_init() {
 	if File_compact_formats_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_compact_formats_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ChainMetadata); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_compact_formats_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CompactBlock); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_compact_formats_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CompactTx); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_compact_formats_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CompactSaplingSpend); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_compact_formats_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CompactSaplingOutput); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_compact_formats_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CompactOrchardAction); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_compact_formats_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_compact_formats_proto_rawDesc), len(file_compact_formats_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   6,
 			NumExtensions: 0,
@@ -678,7 +553,6 @@ func file_compact_formats_proto_init() {
 		MessageInfos:      file_compact_formats_proto_msgTypes,
 	}.Build()
 	File_compact_formats_proto = out.File
-	file_compact_formats_proto_rawDesc = nil
 	file_compact_formats_proto_goTypes = nil
 	file_compact_formats_proto_depIdxs = nil
 }
