@@ -4,8 +4,8 @@
 
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v6.32.0
 // source: darkside.proto
 
 package walletrpc
@@ -19,8 +19,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	DarksideStreamer_Reset_FullMethodName                     = "/cash.z.wallet.sdk.rpc.DarksideStreamer/Reset"
@@ -44,6 +44,11 @@ const (
 // DarksideStreamerClient is the client API for DarksideStreamer service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Darksidewalletd maintains two staging areas, blocks and transactions. The
+// Stage*() gRPCs add items to the staging area; ApplyStaged() "applies" everything
+// in the staging area to the working (operational) state that the mock zcashd
+// serves; transactions are placed into their corresponding blocks (by height).
 type DarksideStreamerClient interface {
 	// Reset reverts all darksidewalletd state (active block range, latest height,
 	// staged blocks and transactions) and lightwalletd state (cache) to empty,
@@ -56,7 +61,7 @@ type DarksideStreamerClient interface {
 	// StageBlocksStream accepts a list of blocks and saves them into the blocks
 	// staging area until ApplyStaged() is called; there is no immediate effect on
 	// the mock zcashd. Blocks are hex-encoded. Order is important, see ApplyStaged.
-	StageBlocksStream(ctx context.Context, opts ...grpc.CallOption) (DarksideStreamer_StageBlocksStreamClient, error)
+	StageBlocksStream(ctx context.Context, opts ...grpc.CallOption) (grpc.ClientStreamingClient[DarksideBlock, Empty], error)
 	// StageBlocks is the same as StageBlocksStream() except the blocks are fetched
 	// from the given URL. Blocks are one per line, hex-encoded (not JSON).
 	StageBlocks(ctx context.Context, in *DarksideBlocksURL, opts ...grpc.CallOption) (*Empty, error)
@@ -71,7 +76,7 @@ type DarksideStreamerClient interface {
 	// are not returned by the production GetTransaction() gRPC until they
 	// appear in a "mined" block (contained in the active blockchain presented
 	// by the mock zcashd).
-	StageTransactionsStream(ctx context.Context, opts ...grpc.CallOption) (DarksideStreamer_StageTransactionsStreamClient, error)
+	StageTransactionsStream(ctx context.Context, opts ...grpc.CallOption) (grpc.ClientStreamingClient[RawTransaction, Empty], error)
 	// StageTransactions is the same except the transactions are fetched from
 	// the given url. They are all staged into the block at the given height.
 	// Staging transactions to different heights requires multiple calls.
@@ -100,7 +105,7 @@ type DarksideStreamerClient interface {
 	// been mined yet. The intention is that the transactions returned here can
 	// then, for example, be given to StageTransactions() to get them "mined"
 	// into a specified block on the next ApplyStaged().
-	GetIncomingTransactions(ctx context.Context, in *Empty, opts ...grpc.CallOption) (DarksideStreamer_GetIncomingTransactionsClient, error)
+	GetIncomingTransactions(ctx context.Context, in *Empty, opts ...grpc.CallOption) (grpc.ServerStreamingClient[RawTransaction], error)
 	// Clear the incoming transaction pool.
 	ClearIncomingTransactions(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Empty, error)
 	// Add a GetAddressUtxosReply entry to be returned by GetAddressUtxos().
@@ -130,51 +135,32 @@ func NewDarksideStreamerClient(cc grpc.ClientConnInterface) DarksideStreamerClie
 }
 
 func (c *darksideStreamerClient) Reset(ctx context.Context, in *DarksideMetaState, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_Reset_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_Reset_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *darksideStreamerClient) StageBlocksStream(ctx context.Context, opts ...grpc.CallOption) (DarksideStreamer_StageBlocksStreamClient, error) {
-	stream, err := c.cc.NewStream(ctx, &DarksideStreamer_ServiceDesc.Streams[0], DarksideStreamer_StageBlocksStream_FullMethodName, opts...)
+func (c *darksideStreamerClient) StageBlocksStream(ctx context.Context, opts ...grpc.CallOption) (grpc.ClientStreamingClient[DarksideBlock, Empty], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &DarksideStreamer_ServiceDesc.Streams[0], DarksideStreamer_StageBlocksStream_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &darksideStreamerStageBlocksStreamClient{stream}
+	x := &grpc.GenericClientStream[DarksideBlock, Empty]{ClientStream: stream}
 	return x, nil
 }
 
-type DarksideStreamer_StageBlocksStreamClient interface {
-	Send(*DarksideBlock) error
-	CloseAndRecv() (*Empty, error)
-	grpc.ClientStream
-}
-
-type darksideStreamerStageBlocksStreamClient struct {
-	grpc.ClientStream
-}
-
-func (x *darksideStreamerStageBlocksStreamClient) Send(m *DarksideBlock) error {
-	return x.ClientStream.SendMsg(m)
-}
-
-func (x *darksideStreamerStageBlocksStreamClient) CloseAndRecv() (*Empty, error) {
-	if err := x.ClientStream.CloseSend(); err != nil {
-		return nil, err
-	}
-	m := new(Empty)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type DarksideStreamer_StageBlocksStreamClient = grpc.ClientStreamingClient[DarksideBlock, Empty]
 
 func (c *darksideStreamerClient) StageBlocks(ctx context.Context, in *DarksideBlocksURL, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_StageBlocks_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_StageBlocks_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -182,51 +168,32 @@ func (c *darksideStreamerClient) StageBlocks(ctx context.Context, in *DarksideBl
 }
 
 func (c *darksideStreamerClient) StageBlocksCreate(ctx context.Context, in *DarksideEmptyBlocks, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_StageBlocksCreate_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_StageBlocksCreate_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *darksideStreamerClient) StageTransactionsStream(ctx context.Context, opts ...grpc.CallOption) (DarksideStreamer_StageTransactionsStreamClient, error) {
-	stream, err := c.cc.NewStream(ctx, &DarksideStreamer_ServiceDesc.Streams[1], DarksideStreamer_StageTransactionsStream_FullMethodName, opts...)
+func (c *darksideStreamerClient) StageTransactionsStream(ctx context.Context, opts ...grpc.CallOption) (grpc.ClientStreamingClient[RawTransaction, Empty], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &DarksideStreamer_ServiceDesc.Streams[1], DarksideStreamer_StageTransactionsStream_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &darksideStreamerStageTransactionsStreamClient{stream}
+	x := &grpc.GenericClientStream[RawTransaction, Empty]{ClientStream: stream}
 	return x, nil
 }
 
-type DarksideStreamer_StageTransactionsStreamClient interface {
-	Send(*RawTransaction) error
-	CloseAndRecv() (*Empty, error)
-	grpc.ClientStream
-}
-
-type darksideStreamerStageTransactionsStreamClient struct {
-	grpc.ClientStream
-}
-
-func (x *darksideStreamerStageTransactionsStreamClient) Send(m *RawTransaction) error {
-	return x.ClientStream.SendMsg(m)
-}
-
-func (x *darksideStreamerStageTransactionsStreamClient) CloseAndRecv() (*Empty, error) {
-	if err := x.ClientStream.CloseSend(); err != nil {
-		return nil, err
-	}
-	m := new(Empty)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type DarksideStreamer_StageTransactionsStreamClient = grpc.ClientStreamingClient[RawTransaction, Empty]
 
 func (c *darksideStreamerClient) StageTransactions(ctx context.Context, in *DarksideTransactionsURL, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_StageTransactions_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_StageTransactions_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -234,20 +201,22 @@ func (c *darksideStreamerClient) StageTransactions(ctx context.Context, in *Dark
 }
 
 func (c *darksideStreamerClient) ApplyStaged(ctx context.Context, in *DarksideHeight, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_ApplyStaged_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_ApplyStaged_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *darksideStreamerClient) GetIncomingTransactions(ctx context.Context, in *Empty, opts ...grpc.CallOption) (DarksideStreamer_GetIncomingTransactionsClient, error) {
-	stream, err := c.cc.NewStream(ctx, &DarksideStreamer_ServiceDesc.Streams[2], DarksideStreamer_GetIncomingTransactions_FullMethodName, opts...)
+func (c *darksideStreamerClient) GetIncomingTransactions(ctx context.Context, in *Empty, opts ...grpc.CallOption) (grpc.ServerStreamingClient[RawTransaction], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &DarksideStreamer_ServiceDesc.Streams[2], DarksideStreamer_GetIncomingTransactions_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &darksideStreamerGetIncomingTransactionsClient{stream}
+	x := &grpc.GenericClientStream[Empty, RawTransaction]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -257,26 +226,13 @@ func (c *darksideStreamerClient) GetIncomingTransactions(ctx context.Context, in
 	return x, nil
 }
 
-type DarksideStreamer_GetIncomingTransactionsClient interface {
-	Recv() (*RawTransaction, error)
-	grpc.ClientStream
-}
-
-type darksideStreamerGetIncomingTransactionsClient struct {
-	grpc.ClientStream
-}
-
-func (x *darksideStreamerGetIncomingTransactionsClient) Recv() (*RawTransaction, error) {
-	m := new(RawTransaction)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type DarksideStreamer_GetIncomingTransactionsClient = grpc.ServerStreamingClient[RawTransaction]
 
 func (c *darksideStreamerClient) ClearIncomingTransactions(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_ClearIncomingTransactions_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_ClearIncomingTransactions_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -284,8 +240,9 @@ func (c *darksideStreamerClient) ClearIncomingTransactions(ctx context.Context, 
 }
 
 func (c *darksideStreamerClient) AddAddressUtxo(ctx context.Context, in *GetAddressUtxosReply, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_AddAddressUtxo_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_AddAddressUtxo_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -293,8 +250,9 @@ func (c *darksideStreamerClient) AddAddressUtxo(ctx context.Context, in *GetAddr
 }
 
 func (c *darksideStreamerClient) ClearAddressUtxo(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_ClearAddressUtxo_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_ClearAddressUtxo_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -302,8 +260,9 @@ func (c *darksideStreamerClient) ClearAddressUtxo(ctx context.Context, in *Empty
 }
 
 func (c *darksideStreamerClient) AddTreeState(ctx context.Context, in *TreeState, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_AddTreeState_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_AddTreeState_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -311,8 +270,9 @@ func (c *darksideStreamerClient) AddTreeState(ctx context.Context, in *TreeState
 }
 
 func (c *darksideStreamerClient) RemoveTreeState(ctx context.Context, in *BlockID, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_RemoveTreeState_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_RemoveTreeState_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -320,8 +280,9 @@ func (c *darksideStreamerClient) RemoveTreeState(ctx context.Context, in *BlockI
 }
 
 func (c *darksideStreamerClient) ClearAllTreeStates(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_ClearAllTreeStates_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_ClearAllTreeStates_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -329,8 +290,9 @@ func (c *darksideStreamerClient) ClearAllTreeStates(ctx context.Context, in *Emp
 }
 
 func (c *darksideStreamerClient) SetSubtreeRoots(ctx context.Context, in *DarksideSubtreeRoots, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_SetSubtreeRoots_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_SetSubtreeRoots_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -338,8 +300,9 @@ func (c *darksideStreamerClient) SetSubtreeRoots(ctx context.Context, in *Darksi
 }
 
 func (c *darksideStreamerClient) Stop(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Empty)
-	err := c.cc.Invoke(ctx, DarksideStreamer_Stop_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, DarksideStreamer_Stop_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -348,7 +311,12 @@ func (c *darksideStreamerClient) Stop(ctx context.Context, in *Empty, opts ...gr
 
 // DarksideStreamerServer is the server API for DarksideStreamer service.
 // All implementations must embed UnimplementedDarksideStreamerServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Darksidewalletd maintains two staging areas, blocks and transactions. The
+// Stage*() gRPCs add items to the staging area; ApplyStaged() "applies" everything
+// in the staging area to the working (operational) state that the mock zcashd
+// serves; transactions are placed into their corresponding blocks (by height).
 type DarksideStreamerServer interface {
 	// Reset reverts all darksidewalletd state (active block range, latest height,
 	// staged blocks and transactions) and lightwalletd state (cache) to empty,
@@ -361,7 +329,7 @@ type DarksideStreamerServer interface {
 	// StageBlocksStream accepts a list of blocks and saves them into the blocks
 	// staging area until ApplyStaged() is called; there is no immediate effect on
 	// the mock zcashd. Blocks are hex-encoded. Order is important, see ApplyStaged.
-	StageBlocksStream(DarksideStreamer_StageBlocksStreamServer) error
+	StageBlocksStream(grpc.ClientStreamingServer[DarksideBlock, Empty]) error
 	// StageBlocks is the same as StageBlocksStream() except the blocks are fetched
 	// from the given URL. Blocks are one per line, hex-encoded (not JSON).
 	StageBlocks(context.Context, *DarksideBlocksURL) (*Empty, error)
@@ -376,7 +344,7 @@ type DarksideStreamerServer interface {
 	// are not returned by the production GetTransaction() gRPC until they
 	// appear in a "mined" block (contained in the active blockchain presented
 	// by the mock zcashd).
-	StageTransactionsStream(DarksideStreamer_StageTransactionsStreamServer) error
+	StageTransactionsStream(grpc.ClientStreamingServer[RawTransaction, Empty]) error
 	// StageTransactions is the same except the transactions are fetched from
 	// the given url. They are all staged into the block at the given height.
 	// Staging transactions to different heights requires multiple calls.
@@ -405,7 +373,7 @@ type DarksideStreamerServer interface {
 	// been mined yet. The intention is that the transactions returned here can
 	// then, for example, be given to StageTransactions() to get them "mined"
 	// into a specified block on the next ApplyStaged().
-	GetIncomingTransactions(*Empty, DarksideStreamer_GetIncomingTransactionsServer) error
+	GetIncomingTransactions(*Empty, grpc.ServerStreamingServer[RawTransaction]) error
 	// Clear the incoming transaction pool.
 	ClearIncomingTransactions(context.Context, *Empty) (*Empty, error)
 	// Add a GetAddressUtxosReply entry to be returned by GetAddressUtxos().
@@ -427,14 +395,17 @@ type DarksideStreamerServer interface {
 	mustEmbedUnimplementedDarksideStreamerServer()
 }
 
-// UnimplementedDarksideStreamerServer must be embedded to have forward compatible implementations.
-type UnimplementedDarksideStreamerServer struct {
-}
+// UnimplementedDarksideStreamerServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedDarksideStreamerServer struct{}
 
 func (UnimplementedDarksideStreamerServer) Reset(context.Context, *DarksideMetaState) (*Empty, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Reset not implemented")
 }
-func (UnimplementedDarksideStreamerServer) StageBlocksStream(DarksideStreamer_StageBlocksStreamServer) error {
+func (UnimplementedDarksideStreamerServer) StageBlocksStream(grpc.ClientStreamingServer[DarksideBlock, Empty]) error {
 	return status.Errorf(codes.Unimplemented, "method StageBlocksStream not implemented")
 }
 func (UnimplementedDarksideStreamerServer) StageBlocks(context.Context, *DarksideBlocksURL) (*Empty, error) {
@@ -443,7 +414,7 @@ func (UnimplementedDarksideStreamerServer) StageBlocks(context.Context, *Darksid
 func (UnimplementedDarksideStreamerServer) StageBlocksCreate(context.Context, *DarksideEmptyBlocks) (*Empty, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method StageBlocksCreate not implemented")
 }
-func (UnimplementedDarksideStreamerServer) StageTransactionsStream(DarksideStreamer_StageTransactionsStreamServer) error {
+func (UnimplementedDarksideStreamerServer) StageTransactionsStream(grpc.ClientStreamingServer[RawTransaction, Empty]) error {
 	return status.Errorf(codes.Unimplemented, "method StageTransactionsStream not implemented")
 }
 func (UnimplementedDarksideStreamerServer) StageTransactions(context.Context, *DarksideTransactionsURL) (*Empty, error) {
@@ -452,7 +423,7 @@ func (UnimplementedDarksideStreamerServer) StageTransactions(context.Context, *D
 func (UnimplementedDarksideStreamerServer) ApplyStaged(context.Context, *DarksideHeight) (*Empty, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ApplyStaged not implemented")
 }
-func (UnimplementedDarksideStreamerServer) GetIncomingTransactions(*Empty, DarksideStreamer_GetIncomingTransactionsServer) error {
+func (UnimplementedDarksideStreamerServer) GetIncomingTransactions(*Empty, grpc.ServerStreamingServer[RawTransaction]) error {
 	return status.Errorf(codes.Unimplemented, "method GetIncomingTransactions not implemented")
 }
 func (UnimplementedDarksideStreamerServer) ClearIncomingTransactions(context.Context, *Empty) (*Empty, error) {
@@ -480,6 +451,7 @@ func (UnimplementedDarksideStreamerServer) Stop(context.Context, *Empty) (*Empty
 	return nil, status.Errorf(codes.Unimplemented, "method Stop not implemented")
 }
 func (UnimplementedDarksideStreamerServer) mustEmbedUnimplementedDarksideStreamerServer() {}
+func (UnimplementedDarksideStreamerServer) testEmbeddedByValue()                          {}
 
 // UnsafeDarksideStreamerServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to DarksideStreamerServer will
@@ -489,6 +461,13 @@ type UnsafeDarksideStreamerServer interface {
 }
 
 func RegisterDarksideStreamerServer(s grpc.ServiceRegistrar, srv DarksideStreamerServer) {
+	// If the following call pancis, it indicates UnimplementedDarksideStreamerServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&DarksideStreamer_ServiceDesc, srv)
 }
 
@@ -511,30 +490,11 @@ func _DarksideStreamer_Reset_Handler(srv interface{}, ctx context.Context, dec f
 }
 
 func _DarksideStreamer_StageBlocksStream_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(DarksideStreamerServer).StageBlocksStream(&darksideStreamerStageBlocksStreamServer{stream})
+	return srv.(DarksideStreamerServer).StageBlocksStream(&grpc.GenericServerStream[DarksideBlock, Empty]{ServerStream: stream})
 }
 
-type DarksideStreamer_StageBlocksStreamServer interface {
-	SendAndClose(*Empty) error
-	Recv() (*DarksideBlock, error)
-	grpc.ServerStream
-}
-
-type darksideStreamerStageBlocksStreamServer struct {
-	grpc.ServerStream
-}
-
-func (x *darksideStreamerStageBlocksStreamServer) SendAndClose(m *Empty) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-func (x *darksideStreamerStageBlocksStreamServer) Recv() (*DarksideBlock, error) {
-	m := new(DarksideBlock)
-	if err := x.ServerStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type DarksideStreamer_StageBlocksStreamServer = grpc.ClientStreamingServer[DarksideBlock, Empty]
 
 func _DarksideStreamer_StageBlocks_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(DarksideBlocksURL)
@@ -573,30 +533,11 @@ func _DarksideStreamer_StageBlocksCreate_Handler(srv interface{}, ctx context.Co
 }
 
 func _DarksideStreamer_StageTransactionsStream_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(DarksideStreamerServer).StageTransactionsStream(&darksideStreamerStageTransactionsStreamServer{stream})
+	return srv.(DarksideStreamerServer).StageTransactionsStream(&grpc.GenericServerStream[RawTransaction, Empty]{ServerStream: stream})
 }
 
-type DarksideStreamer_StageTransactionsStreamServer interface {
-	SendAndClose(*Empty) error
-	Recv() (*RawTransaction, error)
-	grpc.ServerStream
-}
-
-type darksideStreamerStageTransactionsStreamServer struct {
-	grpc.ServerStream
-}
-
-func (x *darksideStreamerStageTransactionsStreamServer) SendAndClose(m *Empty) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-func (x *darksideStreamerStageTransactionsStreamServer) Recv() (*RawTransaction, error) {
-	m := new(RawTransaction)
-	if err := x.ServerStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type DarksideStreamer_StageTransactionsStreamServer = grpc.ClientStreamingServer[RawTransaction, Empty]
 
 func _DarksideStreamer_StageTransactions_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(DarksideTransactionsURL)
@@ -639,21 +580,11 @@ func _DarksideStreamer_GetIncomingTransactions_Handler(srv interface{}, stream g
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(DarksideStreamerServer).GetIncomingTransactions(m, &darksideStreamerGetIncomingTransactionsServer{stream})
+	return srv.(DarksideStreamerServer).GetIncomingTransactions(m, &grpc.GenericServerStream[Empty, RawTransaction]{ServerStream: stream})
 }
 
-type DarksideStreamer_GetIncomingTransactionsServer interface {
-	Send(*RawTransaction) error
-	grpc.ServerStream
-}
-
-type darksideStreamerGetIncomingTransactionsServer struct {
-	grpc.ServerStream
-}
-
-func (x *darksideStreamerGetIncomingTransactionsServer) Send(m *RawTransaction) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type DarksideStreamer_GetIncomingTransactionsServer = grpc.ServerStreamingServer[RawTransaction]
 
 func _DarksideStreamer_ClearIncomingTransactions_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(Empty)
