@@ -4,8 +4,8 @@
 
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.25.2
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v6.32.0
 // source: service.proto
 
 package walletrpc
@@ -19,8 +19,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	CompactTxStreamer_GetLatestBlock_FullMethodName           = "/cash.z.wallet.sdk.rpc.CompactTxStreamer/GetLatestBlock"
@@ -56,9 +56,9 @@ type CompactTxStreamerClient interface {
 	// Same as GetBlock except actions contain only nullifiers
 	GetBlockNullifiers(ctx context.Context, in *BlockID, opts ...grpc.CallOption) (*CompactBlock, error)
 	// Return a list of consecutive compact blocks
-	GetBlockRange(ctx context.Context, in *BlockRange, opts ...grpc.CallOption) (CompactTxStreamer_GetBlockRangeClient, error)
+	GetBlockRange(ctx context.Context, in *BlockRange, opts ...grpc.CallOption) (grpc.ServerStreamingClient[CompactBlock], error)
 	// Same as GetBlockRange except actions contain only nullifiers
-	GetBlockRangeNullifiers(ctx context.Context, in *BlockRange, opts ...grpc.CallOption) (CompactTxStreamer_GetBlockRangeNullifiersClient, error)
+	GetBlockRangeNullifiers(ctx context.Context, in *BlockRange, opts ...grpc.CallOption) (grpc.ServerStreamingClient[CompactBlock], error)
 	// Return the requested full (not compact) transaction (as from zcashd)
 	GetTransaction(ctx context.Context, in *TxFilter, opts ...grpc.CallOption) (*RawTransaction, error)
 	// Submit the given transaction to the Zcash network
@@ -66,11 +66,11 @@ type CompactTxStreamerClient interface {
 	// Return the transactions corresponding to the given t-address within the given block range
 	// NB - this method is misnamed, it returns transactions, not transaction IDs.
 	// NOTE: this method is deprecated, please use GetTaddressTransactions instead.
-	GetTaddressTxids(ctx context.Context, in *TransparentAddressBlockFilter, opts ...grpc.CallOption) (CompactTxStreamer_GetTaddressTxidsClient, error)
+	GetTaddressTxids(ctx context.Context, in *TransparentAddressBlockFilter, opts ...grpc.CallOption) (grpc.ServerStreamingClient[RawTransaction], error)
 	// Return the transactions corresponding to the given t-address within the given block range
-	GetTaddressTransactions(ctx context.Context, in *TransparentAddressBlockFilter, opts ...grpc.CallOption) (CompactTxStreamer_GetTaddressTransactionsClient, error)
+	GetTaddressTransactions(ctx context.Context, in *TransparentAddressBlockFilter, opts ...grpc.CallOption) (grpc.ServerStreamingClient[RawTransaction], error)
 	GetTaddressBalance(ctx context.Context, in *AddressList, opts ...grpc.CallOption) (*Balance, error)
-	GetTaddressBalanceStream(ctx context.Context, opts ...grpc.CallOption) (CompactTxStreamer_GetTaddressBalanceStreamClient, error)
+	GetTaddressBalanceStream(ctx context.Context, opts ...grpc.CallOption) (grpc.ClientStreamingClient[Address, Balance], error)
 	// Return the compact transactions currently in the mempool; the results
 	// can be a few seconds out of date. If the Exclude list is empty, return
 	// all transactions; otherwise return all *except* those in the Exclude list
@@ -83,10 +83,10 @@ type CompactTxStreamerClient interface {
 	//
 	// The a shortened transaction ID is the prefix in big-endian (hex) format
 	// (then converted to binary). See smoke-test.bash for examples.
-	GetMempoolTx(ctx context.Context, in *Exclude, opts ...grpc.CallOption) (CompactTxStreamer_GetMempoolTxClient, error)
+	GetMempoolTx(ctx context.Context, in *Exclude, opts ...grpc.CallOption) (grpc.ServerStreamingClient[CompactTx], error)
 	// Return a stream of current Mempool transactions. This will keep the output stream open while
 	// there are mempool transactions. It will close the returned stream when a new block is mined.
-	GetMempoolStream(ctx context.Context, in *Empty, opts ...grpc.CallOption) (CompactTxStreamer_GetMempoolStreamClient, error)
+	GetMempoolStream(ctx context.Context, in *Empty, opts ...grpc.CallOption) (grpc.ServerStreamingClient[RawTransaction], error)
 	// GetTreeState returns the note commitment tree state corresponding to the given block.
 	// See section 3.7 of the Zcash protocol specification. It returns several other useful
 	// values also (even though they can be obtained using GetBlock).
@@ -95,9 +95,9 @@ type CompactTxStreamerClient interface {
 	GetLatestTreeState(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*TreeState, error)
 	// Returns a stream of information about roots of subtrees of the note commitment tree
 	// for the specified shielded protocol (Sapling or Orchard).
-	GetSubtreeRoots(ctx context.Context, in *GetSubtreeRootsArg, opts ...grpc.CallOption) (CompactTxStreamer_GetSubtreeRootsClient, error)
+	GetSubtreeRoots(ctx context.Context, in *GetSubtreeRootsArg, opts ...grpc.CallOption) (grpc.ServerStreamingClient[SubtreeRoot], error)
 	GetAddressUtxos(ctx context.Context, in *GetAddressUtxosArg, opts ...grpc.CallOption) (*GetAddressUtxosReplyList, error)
-	GetAddressUtxosStream(ctx context.Context, in *GetAddressUtxosArg, opts ...grpc.CallOption) (CompactTxStreamer_GetAddressUtxosStreamClient, error)
+	GetAddressUtxosStream(ctx context.Context, in *GetAddressUtxosArg, opts ...grpc.CallOption) (grpc.ServerStreamingClient[GetAddressUtxosReply], error)
 	// Return information about this lightwalletd instance and the blockchain
 	GetLightdInfo(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*LightdInfo, error)
 	// Testing-only, requires lightwalletd --ping-very-insecure (do not enable in production)
@@ -113,8 +113,9 @@ func NewCompactTxStreamerClient(cc grpc.ClientConnInterface) CompactTxStreamerCl
 }
 
 func (c *compactTxStreamerClient) GetLatestBlock(ctx context.Context, in *ChainSpec, opts ...grpc.CallOption) (*BlockID, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(BlockID)
-	err := c.cc.Invoke(ctx, CompactTxStreamer_GetLatestBlock_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CompactTxStreamer_GetLatestBlock_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -122,8 +123,9 @@ func (c *compactTxStreamerClient) GetLatestBlock(ctx context.Context, in *ChainS
 }
 
 func (c *compactTxStreamerClient) GetBlock(ctx context.Context, in *BlockID, opts ...grpc.CallOption) (*CompactBlock, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CompactBlock)
-	err := c.cc.Invoke(ctx, CompactTxStreamer_GetBlock_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CompactTxStreamer_GetBlock_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -131,20 +133,22 @@ func (c *compactTxStreamerClient) GetBlock(ctx context.Context, in *BlockID, opt
 }
 
 func (c *compactTxStreamerClient) GetBlockNullifiers(ctx context.Context, in *BlockID, opts ...grpc.CallOption) (*CompactBlock, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CompactBlock)
-	err := c.cc.Invoke(ctx, CompactTxStreamer_GetBlockNullifiers_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CompactTxStreamer_GetBlockNullifiers_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *compactTxStreamerClient) GetBlockRange(ctx context.Context, in *BlockRange, opts ...grpc.CallOption) (CompactTxStreamer_GetBlockRangeClient, error) {
-	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[0], CompactTxStreamer_GetBlockRange_FullMethodName, opts...)
+func (c *compactTxStreamerClient) GetBlockRange(ctx context.Context, in *BlockRange, opts ...grpc.CallOption) (grpc.ServerStreamingClient[CompactBlock], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[0], CompactTxStreamer_GetBlockRange_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &compactTxStreamerGetBlockRangeClient{stream}
+	x := &grpc.GenericClientStream[BlockRange, CompactBlock]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -154,29 +158,16 @@ func (c *compactTxStreamerClient) GetBlockRange(ctx context.Context, in *BlockRa
 	return x, nil
 }
 
-type CompactTxStreamer_GetBlockRangeClient interface {
-	Recv() (*CompactBlock, error)
-	grpc.ClientStream
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetBlockRangeClient = grpc.ServerStreamingClient[CompactBlock]
 
-type compactTxStreamerGetBlockRangeClient struct {
-	grpc.ClientStream
-}
-
-func (x *compactTxStreamerGetBlockRangeClient) Recv() (*CompactBlock, error) {
-	m := new(CompactBlock)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func (c *compactTxStreamerClient) GetBlockRangeNullifiers(ctx context.Context, in *BlockRange, opts ...grpc.CallOption) (CompactTxStreamer_GetBlockRangeNullifiersClient, error) {
-	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[1], CompactTxStreamer_GetBlockRangeNullifiers_FullMethodName, opts...)
+func (c *compactTxStreamerClient) GetBlockRangeNullifiers(ctx context.Context, in *BlockRange, opts ...grpc.CallOption) (grpc.ServerStreamingClient[CompactBlock], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[1], CompactTxStreamer_GetBlockRangeNullifiers_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &compactTxStreamerGetBlockRangeNullifiersClient{stream}
+	x := &grpc.GenericClientStream[BlockRange, CompactBlock]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -186,26 +177,13 @@ func (c *compactTxStreamerClient) GetBlockRangeNullifiers(ctx context.Context, i
 	return x, nil
 }
 
-type CompactTxStreamer_GetBlockRangeNullifiersClient interface {
-	Recv() (*CompactBlock, error)
-	grpc.ClientStream
-}
-
-type compactTxStreamerGetBlockRangeNullifiersClient struct {
-	grpc.ClientStream
-}
-
-func (x *compactTxStreamerGetBlockRangeNullifiersClient) Recv() (*CompactBlock, error) {
-	m := new(CompactBlock)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetBlockRangeNullifiersClient = grpc.ServerStreamingClient[CompactBlock]
 
 func (c *compactTxStreamerClient) GetTransaction(ctx context.Context, in *TxFilter, opts ...grpc.CallOption) (*RawTransaction, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(RawTransaction)
-	err := c.cc.Invoke(ctx, CompactTxStreamer_GetTransaction_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CompactTxStreamer_GetTransaction_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -213,20 +191,22 @@ func (c *compactTxStreamerClient) GetTransaction(ctx context.Context, in *TxFilt
 }
 
 func (c *compactTxStreamerClient) SendTransaction(ctx context.Context, in *RawTransaction, opts ...grpc.CallOption) (*SendResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(SendResponse)
-	err := c.cc.Invoke(ctx, CompactTxStreamer_SendTransaction_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CompactTxStreamer_SendTransaction_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *compactTxStreamerClient) GetTaddressTxids(ctx context.Context, in *TransparentAddressBlockFilter, opts ...grpc.CallOption) (CompactTxStreamer_GetTaddressTxidsClient, error) {
-	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[2], CompactTxStreamer_GetTaddressTxids_FullMethodName, opts...)
+func (c *compactTxStreamerClient) GetTaddressTxids(ctx context.Context, in *TransparentAddressBlockFilter, opts ...grpc.CallOption) (grpc.ServerStreamingClient[RawTransaction], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[2], CompactTxStreamer_GetTaddressTxids_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &compactTxStreamerGetTaddressTxidsClient{stream}
+	x := &grpc.GenericClientStream[TransparentAddressBlockFilter, RawTransaction]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -236,29 +216,16 @@ func (c *compactTxStreamerClient) GetTaddressTxids(ctx context.Context, in *Tran
 	return x, nil
 }
 
-type CompactTxStreamer_GetTaddressTxidsClient interface {
-	Recv() (*RawTransaction, error)
-	grpc.ClientStream
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetTaddressTxidsClient = grpc.ServerStreamingClient[RawTransaction]
 
-type compactTxStreamerGetTaddressTxidsClient struct {
-	grpc.ClientStream
-}
-
-func (x *compactTxStreamerGetTaddressTxidsClient) Recv() (*RawTransaction, error) {
-	m := new(RawTransaction)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func (c *compactTxStreamerClient) GetTaddressTransactions(ctx context.Context, in *TransparentAddressBlockFilter, opts ...grpc.CallOption) (CompactTxStreamer_GetTaddressTransactionsClient, error) {
-	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[3], CompactTxStreamer_GetTaddressTransactions_FullMethodName, opts...)
+func (c *compactTxStreamerClient) GetTaddressTransactions(ctx context.Context, in *TransparentAddressBlockFilter, opts ...grpc.CallOption) (grpc.ServerStreamingClient[RawTransaction], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[3], CompactTxStreamer_GetTaddressTransactions_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &compactTxStreamerGetTaddressTransactionsClient{stream}
+	x := &grpc.GenericClientStream[TransparentAddressBlockFilter, RawTransaction]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -268,72 +235,39 @@ func (c *compactTxStreamerClient) GetTaddressTransactions(ctx context.Context, i
 	return x, nil
 }
 
-type CompactTxStreamer_GetTaddressTransactionsClient interface {
-	Recv() (*RawTransaction, error)
-	grpc.ClientStream
-}
-
-type compactTxStreamerGetTaddressTransactionsClient struct {
-	grpc.ClientStream
-}
-
-func (x *compactTxStreamerGetTaddressTransactionsClient) Recv() (*RawTransaction, error) {
-	m := new(RawTransaction)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetTaddressTransactionsClient = grpc.ServerStreamingClient[RawTransaction]
 
 func (c *compactTxStreamerClient) GetTaddressBalance(ctx context.Context, in *AddressList, opts ...grpc.CallOption) (*Balance, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Balance)
-	err := c.cc.Invoke(ctx, CompactTxStreamer_GetTaddressBalance_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CompactTxStreamer_GetTaddressBalance_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *compactTxStreamerClient) GetTaddressBalanceStream(ctx context.Context, opts ...grpc.CallOption) (CompactTxStreamer_GetTaddressBalanceStreamClient, error) {
-	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[4], CompactTxStreamer_GetTaddressBalanceStream_FullMethodName, opts...)
+func (c *compactTxStreamerClient) GetTaddressBalanceStream(ctx context.Context, opts ...grpc.CallOption) (grpc.ClientStreamingClient[Address, Balance], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[4], CompactTxStreamer_GetTaddressBalanceStream_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &compactTxStreamerGetTaddressBalanceStreamClient{stream}
+	x := &grpc.GenericClientStream[Address, Balance]{ClientStream: stream}
 	return x, nil
 }
 
-type CompactTxStreamer_GetTaddressBalanceStreamClient interface {
-	Send(*Address) error
-	CloseAndRecv() (*Balance, error)
-	grpc.ClientStream
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetTaddressBalanceStreamClient = grpc.ClientStreamingClient[Address, Balance]
 
-type compactTxStreamerGetTaddressBalanceStreamClient struct {
-	grpc.ClientStream
-}
-
-func (x *compactTxStreamerGetTaddressBalanceStreamClient) Send(m *Address) error {
-	return x.ClientStream.SendMsg(m)
-}
-
-func (x *compactTxStreamerGetTaddressBalanceStreamClient) CloseAndRecv() (*Balance, error) {
-	if err := x.ClientStream.CloseSend(); err != nil {
-		return nil, err
-	}
-	m := new(Balance)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func (c *compactTxStreamerClient) GetMempoolTx(ctx context.Context, in *Exclude, opts ...grpc.CallOption) (CompactTxStreamer_GetMempoolTxClient, error) {
-	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[5], CompactTxStreamer_GetMempoolTx_FullMethodName, opts...)
+func (c *compactTxStreamerClient) GetMempoolTx(ctx context.Context, in *Exclude, opts ...grpc.CallOption) (grpc.ServerStreamingClient[CompactTx], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[5], CompactTxStreamer_GetMempoolTx_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &compactTxStreamerGetMempoolTxClient{stream}
+	x := &grpc.GenericClientStream[Exclude, CompactTx]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -343,29 +277,16 @@ func (c *compactTxStreamerClient) GetMempoolTx(ctx context.Context, in *Exclude,
 	return x, nil
 }
 
-type CompactTxStreamer_GetMempoolTxClient interface {
-	Recv() (*CompactTx, error)
-	grpc.ClientStream
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetMempoolTxClient = grpc.ServerStreamingClient[CompactTx]
 
-type compactTxStreamerGetMempoolTxClient struct {
-	grpc.ClientStream
-}
-
-func (x *compactTxStreamerGetMempoolTxClient) Recv() (*CompactTx, error) {
-	m := new(CompactTx)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func (c *compactTxStreamerClient) GetMempoolStream(ctx context.Context, in *Empty, opts ...grpc.CallOption) (CompactTxStreamer_GetMempoolStreamClient, error) {
-	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[6], CompactTxStreamer_GetMempoolStream_FullMethodName, opts...)
+func (c *compactTxStreamerClient) GetMempoolStream(ctx context.Context, in *Empty, opts ...grpc.CallOption) (grpc.ServerStreamingClient[RawTransaction], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[6], CompactTxStreamer_GetMempoolStream_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &compactTxStreamerGetMempoolStreamClient{stream}
+	x := &grpc.GenericClientStream[Empty, RawTransaction]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -375,26 +296,13 @@ func (c *compactTxStreamerClient) GetMempoolStream(ctx context.Context, in *Empt
 	return x, nil
 }
 
-type CompactTxStreamer_GetMempoolStreamClient interface {
-	Recv() (*RawTransaction, error)
-	grpc.ClientStream
-}
-
-type compactTxStreamerGetMempoolStreamClient struct {
-	grpc.ClientStream
-}
-
-func (x *compactTxStreamerGetMempoolStreamClient) Recv() (*RawTransaction, error) {
-	m := new(RawTransaction)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetMempoolStreamClient = grpc.ServerStreamingClient[RawTransaction]
 
 func (c *compactTxStreamerClient) GetTreeState(ctx context.Context, in *BlockID, opts ...grpc.CallOption) (*TreeState, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(TreeState)
-	err := c.cc.Invoke(ctx, CompactTxStreamer_GetTreeState_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CompactTxStreamer_GetTreeState_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -402,20 +310,22 @@ func (c *compactTxStreamerClient) GetTreeState(ctx context.Context, in *BlockID,
 }
 
 func (c *compactTxStreamerClient) GetLatestTreeState(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*TreeState, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(TreeState)
-	err := c.cc.Invoke(ctx, CompactTxStreamer_GetLatestTreeState_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CompactTxStreamer_GetLatestTreeState_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *compactTxStreamerClient) GetSubtreeRoots(ctx context.Context, in *GetSubtreeRootsArg, opts ...grpc.CallOption) (CompactTxStreamer_GetSubtreeRootsClient, error) {
-	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[7], CompactTxStreamer_GetSubtreeRoots_FullMethodName, opts...)
+func (c *compactTxStreamerClient) GetSubtreeRoots(ctx context.Context, in *GetSubtreeRootsArg, opts ...grpc.CallOption) (grpc.ServerStreamingClient[SubtreeRoot], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[7], CompactTxStreamer_GetSubtreeRoots_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &compactTxStreamerGetSubtreeRootsClient{stream}
+	x := &grpc.GenericClientStream[GetSubtreeRootsArg, SubtreeRoot]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -425,38 +335,26 @@ func (c *compactTxStreamerClient) GetSubtreeRoots(ctx context.Context, in *GetSu
 	return x, nil
 }
 
-type CompactTxStreamer_GetSubtreeRootsClient interface {
-	Recv() (*SubtreeRoot, error)
-	grpc.ClientStream
-}
-
-type compactTxStreamerGetSubtreeRootsClient struct {
-	grpc.ClientStream
-}
-
-func (x *compactTxStreamerGetSubtreeRootsClient) Recv() (*SubtreeRoot, error) {
-	m := new(SubtreeRoot)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetSubtreeRootsClient = grpc.ServerStreamingClient[SubtreeRoot]
 
 func (c *compactTxStreamerClient) GetAddressUtxos(ctx context.Context, in *GetAddressUtxosArg, opts ...grpc.CallOption) (*GetAddressUtxosReplyList, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(GetAddressUtxosReplyList)
-	err := c.cc.Invoke(ctx, CompactTxStreamer_GetAddressUtxos_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CompactTxStreamer_GetAddressUtxos_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *compactTxStreamerClient) GetAddressUtxosStream(ctx context.Context, in *GetAddressUtxosArg, opts ...grpc.CallOption) (CompactTxStreamer_GetAddressUtxosStreamClient, error) {
-	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[8], CompactTxStreamer_GetAddressUtxosStream_FullMethodName, opts...)
+func (c *compactTxStreamerClient) GetAddressUtxosStream(ctx context.Context, in *GetAddressUtxosArg, opts ...grpc.CallOption) (grpc.ServerStreamingClient[GetAddressUtxosReply], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &CompactTxStreamer_ServiceDesc.Streams[8], CompactTxStreamer_GetAddressUtxosStream_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &compactTxStreamerGetAddressUtxosStreamClient{stream}
+	x := &grpc.GenericClientStream[GetAddressUtxosArg, GetAddressUtxosReply]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -466,26 +364,13 @@ func (c *compactTxStreamerClient) GetAddressUtxosStream(ctx context.Context, in 
 	return x, nil
 }
 
-type CompactTxStreamer_GetAddressUtxosStreamClient interface {
-	Recv() (*GetAddressUtxosReply, error)
-	grpc.ClientStream
-}
-
-type compactTxStreamerGetAddressUtxosStreamClient struct {
-	grpc.ClientStream
-}
-
-func (x *compactTxStreamerGetAddressUtxosStreamClient) Recv() (*GetAddressUtxosReply, error) {
-	m := new(GetAddressUtxosReply)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetAddressUtxosStreamClient = grpc.ServerStreamingClient[GetAddressUtxosReply]
 
 func (c *compactTxStreamerClient) GetLightdInfo(ctx context.Context, in *Empty, opts ...grpc.CallOption) (*LightdInfo, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(LightdInfo)
-	err := c.cc.Invoke(ctx, CompactTxStreamer_GetLightdInfo_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CompactTxStreamer_GetLightdInfo_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -493,8 +378,9 @@ func (c *compactTxStreamerClient) GetLightdInfo(ctx context.Context, in *Empty, 
 }
 
 func (c *compactTxStreamerClient) Ping(ctx context.Context, in *Duration, opts ...grpc.CallOption) (*PingResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PingResponse)
-	err := c.cc.Invoke(ctx, CompactTxStreamer_Ping_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CompactTxStreamer_Ping_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -503,7 +389,7 @@ func (c *compactTxStreamerClient) Ping(ctx context.Context, in *Duration, opts .
 
 // CompactTxStreamerServer is the server API for CompactTxStreamer service.
 // All implementations must embed UnimplementedCompactTxStreamerServer
-// for forward compatibility
+// for forward compatibility.
 type CompactTxStreamerServer interface {
 	// Return the height of the tip of the best chain
 	GetLatestBlock(context.Context, *ChainSpec) (*BlockID, error)
@@ -512,9 +398,9 @@ type CompactTxStreamerServer interface {
 	// Same as GetBlock except actions contain only nullifiers
 	GetBlockNullifiers(context.Context, *BlockID) (*CompactBlock, error)
 	// Return a list of consecutive compact blocks
-	GetBlockRange(*BlockRange, CompactTxStreamer_GetBlockRangeServer) error
+	GetBlockRange(*BlockRange, grpc.ServerStreamingServer[CompactBlock]) error
 	// Same as GetBlockRange except actions contain only nullifiers
-	GetBlockRangeNullifiers(*BlockRange, CompactTxStreamer_GetBlockRangeNullifiersServer) error
+	GetBlockRangeNullifiers(*BlockRange, grpc.ServerStreamingServer[CompactBlock]) error
 	// Return the requested full (not compact) transaction (as from zcashd)
 	GetTransaction(context.Context, *TxFilter) (*RawTransaction, error)
 	// Submit the given transaction to the Zcash network
@@ -522,11 +408,11 @@ type CompactTxStreamerServer interface {
 	// Return the transactions corresponding to the given t-address within the given block range
 	// NB - this method is misnamed, it returns transactions, not transaction IDs.
 	// NOTE: this method is deprecated, please use GetTaddressTransactions instead.
-	GetTaddressTxids(*TransparentAddressBlockFilter, CompactTxStreamer_GetTaddressTxidsServer) error
+	GetTaddressTxids(*TransparentAddressBlockFilter, grpc.ServerStreamingServer[RawTransaction]) error
 	// Return the transactions corresponding to the given t-address within the given block range
-	GetTaddressTransactions(*TransparentAddressBlockFilter, CompactTxStreamer_GetTaddressTransactionsServer) error
+	GetTaddressTransactions(*TransparentAddressBlockFilter, grpc.ServerStreamingServer[RawTransaction]) error
 	GetTaddressBalance(context.Context, *AddressList) (*Balance, error)
-	GetTaddressBalanceStream(CompactTxStreamer_GetTaddressBalanceStreamServer) error
+	GetTaddressBalanceStream(grpc.ClientStreamingServer[Address, Balance]) error
 	// Return the compact transactions currently in the mempool; the results
 	// can be a few seconds out of date. If the Exclude list is empty, return
 	// all transactions; otherwise return all *except* those in the Exclude list
@@ -539,10 +425,10 @@ type CompactTxStreamerServer interface {
 	//
 	// The a shortened transaction ID is the prefix in big-endian (hex) format
 	// (then converted to binary). See smoke-test.bash for examples.
-	GetMempoolTx(*Exclude, CompactTxStreamer_GetMempoolTxServer) error
+	GetMempoolTx(*Exclude, grpc.ServerStreamingServer[CompactTx]) error
 	// Return a stream of current Mempool transactions. This will keep the output stream open while
 	// there are mempool transactions. It will close the returned stream when a new block is mined.
-	GetMempoolStream(*Empty, CompactTxStreamer_GetMempoolStreamServer) error
+	GetMempoolStream(*Empty, grpc.ServerStreamingServer[RawTransaction]) error
 	// GetTreeState returns the note commitment tree state corresponding to the given block.
 	// See section 3.7 of the Zcash protocol specification. It returns several other useful
 	// values also (even though they can be obtained using GetBlock).
@@ -551,9 +437,9 @@ type CompactTxStreamerServer interface {
 	GetLatestTreeState(context.Context, *Empty) (*TreeState, error)
 	// Returns a stream of information about roots of subtrees of the note commitment tree
 	// for the specified shielded protocol (Sapling or Orchard).
-	GetSubtreeRoots(*GetSubtreeRootsArg, CompactTxStreamer_GetSubtreeRootsServer) error
+	GetSubtreeRoots(*GetSubtreeRootsArg, grpc.ServerStreamingServer[SubtreeRoot]) error
 	GetAddressUtxos(context.Context, *GetAddressUtxosArg) (*GetAddressUtxosReplyList, error)
-	GetAddressUtxosStream(*GetAddressUtxosArg, CompactTxStreamer_GetAddressUtxosStreamServer) error
+	GetAddressUtxosStream(*GetAddressUtxosArg, grpc.ServerStreamingServer[GetAddressUtxosReply]) error
 	// Return information about this lightwalletd instance and the blockchain
 	GetLightdInfo(context.Context, *Empty) (*LightdInfo, error)
 	// Testing-only, requires lightwalletd --ping-very-insecure (do not enable in production)
@@ -561,9 +447,12 @@ type CompactTxStreamerServer interface {
 	mustEmbedUnimplementedCompactTxStreamerServer()
 }
 
-// UnimplementedCompactTxStreamerServer must be embedded to have forward compatible implementations.
-type UnimplementedCompactTxStreamerServer struct {
-}
+// UnimplementedCompactTxStreamerServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedCompactTxStreamerServer struct{}
 
 func (UnimplementedCompactTxStreamerServer) GetLatestBlock(context.Context, *ChainSpec) (*BlockID, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetLatestBlock not implemented")
@@ -574,10 +463,10 @@ func (UnimplementedCompactTxStreamerServer) GetBlock(context.Context, *BlockID) 
 func (UnimplementedCompactTxStreamerServer) GetBlockNullifiers(context.Context, *BlockID) (*CompactBlock, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetBlockNullifiers not implemented")
 }
-func (UnimplementedCompactTxStreamerServer) GetBlockRange(*BlockRange, CompactTxStreamer_GetBlockRangeServer) error {
+func (UnimplementedCompactTxStreamerServer) GetBlockRange(*BlockRange, grpc.ServerStreamingServer[CompactBlock]) error {
 	return status.Errorf(codes.Unimplemented, "method GetBlockRange not implemented")
 }
-func (UnimplementedCompactTxStreamerServer) GetBlockRangeNullifiers(*BlockRange, CompactTxStreamer_GetBlockRangeNullifiersServer) error {
+func (UnimplementedCompactTxStreamerServer) GetBlockRangeNullifiers(*BlockRange, grpc.ServerStreamingServer[CompactBlock]) error {
 	return status.Errorf(codes.Unimplemented, "method GetBlockRangeNullifiers not implemented")
 }
 func (UnimplementedCompactTxStreamerServer) GetTransaction(context.Context, *TxFilter) (*RawTransaction, error) {
@@ -586,22 +475,22 @@ func (UnimplementedCompactTxStreamerServer) GetTransaction(context.Context, *TxF
 func (UnimplementedCompactTxStreamerServer) SendTransaction(context.Context, *RawTransaction) (*SendResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method SendTransaction not implemented")
 }
-func (UnimplementedCompactTxStreamerServer) GetTaddressTxids(*TransparentAddressBlockFilter, CompactTxStreamer_GetTaddressTxidsServer) error {
+func (UnimplementedCompactTxStreamerServer) GetTaddressTxids(*TransparentAddressBlockFilter, grpc.ServerStreamingServer[RawTransaction]) error {
 	return status.Errorf(codes.Unimplemented, "method GetTaddressTxids not implemented")
 }
-func (UnimplementedCompactTxStreamerServer) GetTaddressTransactions(*TransparentAddressBlockFilter, CompactTxStreamer_GetTaddressTransactionsServer) error {
+func (UnimplementedCompactTxStreamerServer) GetTaddressTransactions(*TransparentAddressBlockFilter, grpc.ServerStreamingServer[RawTransaction]) error {
 	return status.Errorf(codes.Unimplemented, "method GetTaddressTransactions not implemented")
 }
 func (UnimplementedCompactTxStreamerServer) GetTaddressBalance(context.Context, *AddressList) (*Balance, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetTaddressBalance not implemented")
 }
-func (UnimplementedCompactTxStreamerServer) GetTaddressBalanceStream(CompactTxStreamer_GetTaddressBalanceStreamServer) error {
+func (UnimplementedCompactTxStreamerServer) GetTaddressBalanceStream(grpc.ClientStreamingServer[Address, Balance]) error {
 	return status.Errorf(codes.Unimplemented, "method GetTaddressBalanceStream not implemented")
 }
-func (UnimplementedCompactTxStreamerServer) GetMempoolTx(*Exclude, CompactTxStreamer_GetMempoolTxServer) error {
+func (UnimplementedCompactTxStreamerServer) GetMempoolTx(*Exclude, grpc.ServerStreamingServer[CompactTx]) error {
 	return status.Errorf(codes.Unimplemented, "method GetMempoolTx not implemented")
 }
-func (UnimplementedCompactTxStreamerServer) GetMempoolStream(*Empty, CompactTxStreamer_GetMempoolStreamServer) error {
+func (UnimplementedCompactTxStreamerServer) GetMempoolStream(*Empty, grpc.ServerStreamingServer[RawTransaction]) error {
 	return status.Errorf(codes.Unimplemented, "method GetMempoolStream not implemented")
 }
 func (UnimplementedCompactTxStreamerServer) GetTreeState(context.Context, *BlockID) (*TreeState, error) {
@@ -610,13 +499,13 @@ func (UnimplementedCompactTxStreamerServer) GetTreeState(context.Context, *Block
 func (UnimplementedCompactTxStreamerServer) GetLatestTreeState(context.Context, *Empty) (*TreeState, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetLatestTreeState not implemented")
 }
-func (UnimplementedCompactTxStreamerServer) GetSubtreeRoots(*GetSubtreeRootsArg, CompactTxStreamer_GetSubtreeRootsServer) error {
+func (UnimplementedCompactTxStreamerServer) GetSubtreeRoots(*GetSubtreeRootsArg, grpc.ServerStreamingServer[SubtreeRoot]) error {
 	return status.Errorf(codes.Unimplemented, "method GetSubtreeRoots not implemented")
 }
 func (UnimplementedCompactTxStreamerServer) GetAddressUtxos(context.Context, *GetAddressUtxosArg) (*GetAddressUtxosReplyList, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetAddressUtxos not implemented")
 }
-func (UnimplementedCompactTxStreamerServer) GetAddressUtxosStream(*GetAddressUtxosArg, CompactTxStreamer_GetAddressUtxosStreamServer) error {
+func (UnimplementedCompactTxStreamerServer) GetAddressUtxosStream(*GetAddressUtxosArg, grpc.ServerStreamingServer[GetAddressUtxosReply]) error {
 	return status.Errorf(codes.Unimplemented, "method GetAddressUtxosStream not implemented")
 }
 func (UnimplementedCompactTxStreamerServer) GetLightdInfo(context.Context, *Empty) (*LightdInfo, error) {
@@ -626,6 +515,7 @@ func (UnimplementedCompactTxStreamerServer) Ping(context.Context, *Duration) (*P
 	return nil, status.Errorf(codes.Unimplemented, "method Ping not implemented")
 }
 func (UnimplementedCompactTxStreamerServer) mustEmbedUnimplementedCompactTxStreamerServer() {}
+func (UnimplementedCompactTxStreamerServer) testEmbeddedByValue()                           {}
 
 // UnsafeCompactTxStreamerServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to CompactTxStreamerServer will
@@ -635,6 +525,13 @@ type UnsafeCompactTxStreamerServer interface {
 }
 
 func RegisterCompactTxStreamerServer(s grpc.ServiceRegistrar, srv CompactTxStreamerServer) {
+	// If the following call pancis, it indicates UnimplementedCompactTxStreamerServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&CompactTxStreamer_ServiceDesc, srv)
 }
 
@@ -697,42 +594,22 @@ func _CompactTxStreamer_GetBlockRange_Handler(srv interface{}, stream grpc.Serve
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(CompactTxStreamerServer).GetBlockRange(m, &compactTxStreamerGetBlockRangeServer{stream})
+	return srv.(CompactTxStreamerServer).GetBlockRange(m, &grpc.GenericServerStream[BlockRange, CompactBlock]{ServerStream: stream})
 }
 
-type CompactTxStreamer_GetBlockRangeServer interface {
-	Send(*CompactBlock) error
-	grpc.ServerStream
-}
-
-type compactTxStreamerGetBlockRangeServer struct {
-	grpc.ServerStream
-}
-
-func (x *compactTxStreamerGetBlockRangeServer) Send(m *CompactBlock) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetBlockRangeServer = grpc.ServerStreamingServer[CompactBlock]
 
 func _CompactTxStreamer_GetBlockRangeNullifiers_Handler(srv interface{}, stream grpc.ServerStream) error {
 	m := new(BlockRange)
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(CompactTxStreamerServer).GetBlockRangeNullifiers(m, &compactTxStreamerGetBlockRangeNullifiersServer{stream})
+	return srv.(CompactTxStreamerServer).GetBlockRangeNullifiers(m, &grpc.GenericServerStream[BlockRange, CompactBlock]{ServerStream: stream})
 }
 
-type CompactTxStreamer_GetBlockRangeNullifiersServer interface {
-	Send(*CompactBlock) error
-	grpc.ServerStream
-}
-
-type compactTxStreamerGetBlockRangeNullifiersServer struct {
-	grpc.ServerStream
-}
-
-func (x *compactTxStreamerGetBlockRangeNullifiersServer) Send(m *CompactBlock) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetBlockRangeNullifiersServer = grpc.ServerStreamingServer[CompactBlock]
 
 func _CompactTxStreamer_GetTransaction_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(TxFilter)
@@ -775,42 +652,22 @@ func _CompactTxStreamer_GetTaddressTxids_Handler(srv interface{}, stream grpc.Se
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(CompactTxStreamerServer).GetTaddressTxids(m, &compactTxStreamerGetTaddressTxidsServer{stream})
+	return srv.(CompactTxStreamerServer).GetTaddressTxids(m, &grpc.GenericServerStream[TransparentAddressBlockFilter, RawTransaction]{ServerStream: stream})
 }
 
-type CompactTxStreamer_GetTaddressTxidsServer interface {
-	Send(*RawTransaction) error
-	grpc.ServerStream
-}
-
-type compactTxStreamerGetTaddressTxidsServer struct {
-	grpc.ServerStream
-}
-
-func (x *compactTxStreamerGetTaddressTxidsServer) Send(m *RawTransaction) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetTaddressTxidsServer = grpc.ServerStreamingServer[RawTransaction]
 
 func _CompactTxStreamer_GetTaddressTransactions_Handler(srv interface{}, stream grpc.ServerStream) error {
 	m := new(TransparentAddressBlockFilter)
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(CompactTxStreamerServer).GetTaddressTransactions(m, &compactTxStreamerGetTaddressTransactionsServer{stream})
+	return srv.(CompactTxStreamerServer).GetTaddressTransactions(m, &grpc.GenericServerStream[TransparentAddressBlockFilter, RawTransaction]{ServerStream: stream})
 }
 
-type CompactTxStreamer_GetTaddressTransactionsServer interface {
-	Send(*RawTransaction) error
-	grpc.ServerStream
-}
-
-type compactTxStreamerGetTaddressTransactionsServer struct {
-	grpc.ServerStream
-}
-
-func (x *compactTxStreamerGetTaddressTransactionsServer) Send(m *RawTransaction) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetTaddressTransactionsServer = grpc.ServerStreamingServer[RawTransaction]
 
 func _CompactTxStreamer_GetTaddressBalance_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(AddressList)
@@ -831,72 +688,33 @@ func _CompactTxStreamer_GetTaddressBalance_Handler(srv interface{}, ctx context.
 }
 
 func _CompactTxStreamer_GetTaddressBalanceStream_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(CompactTxStreamerServer).GetTaddressBalanceStream(&compactTxStreamerGetTaddressBalanceStreamServer{stream})
+	return srv.(CompactTxStreamerServer).GetTaddressBalanceStream(&grpc.GenericServerStream[Address, Balance]{ServerStream: stream})
 }
 
-type CompactTxStreamer_GetTaddressBalanceStreamServer interface {
-	SendAndClose(*Balance) error
-	Recv() (*Address, error)
-	grpc.ServerStream
-}
-
-type compactTxStreamerGetTaddressBalanceStreamServer struct {
-	grpc.ServerStream
-}
-
-func (x *compactTxStreamerGetTaddressBalanceStreamServer) SendAndClose(m *Balance) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-func (x *compactTxStreamerGetTaddressBalanceStreamServer) Recv() (*Address, error) {
-	m := new(Address)
-	if err := x.ServerStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetTaddressBalanceStreamServer = grpc.ClientStreamingServer[Address, Balance]
 
 func _CompactTxStreamer_GetMempoolTx_Handler(srv interface{}, stream grpc.ServerStream) error {
 	m := new(Exclude)
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(CompactTxStreamerServer).GetMempoolTx(m, &compactTxStreamerGetMempoolTxServer{stream})
+	return srv.(CompactTxStreamerServer).GetMempoolTx(m, &grpc.GenericServerStream[Exclude, CompactTx]{ServerStream: stream})
 }
 
-type CompactTxStreamer_GetMempoolTxServer interface {
-	Send(*CompactTx) error
-	grpc.ServerStream
-}
-
-type compactTxStreamerGetMempoolTxServer struct {
-	grpc.ServerStream
-}
-
-func (x *compactTxStreamerGetMempoolTxServer) Send(m *CompactTx) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetMempoolTxServer = grpc.ServerStreamingServer[CompactTx]
 
 func _CompactTxStreamer_GetMempoolStream_Handler(srv interface{}, stream grpc.ServerStream) error {
 	m := new(Empty)
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(CompactTxStreamerServer).GetMempoolStream(m, &compactTxStreamerGetMempoolStreamServer{stream})
+	return srv.(CompactTxStreamerServer).GetMempoolStream(m, &grpc.GenericServerStream[Empty, RawTransaction]{ServerStream: stream})
 }
 
-type CompactTxStreamer_GetMempoolStreamServer interface {
-	Send(*RawTransaction) error
-	grpc.ServerStream
-}
-
-type compactTxStreamerGetMempoolStreamServer struct {
-	grpc.ServerStream
-}
-
-func (x *compactTxStreamerGetMempoolStreamServer) Send(m *RawTransaction) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetMempoolStreamServer = grpc.ServerStreamingServer[RawTransaction]
 
 func _CompactTxStreamer_GetTreeState_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(BlockID)
@@ -939,21 +757,11 @@ func _CompactTxStreamer_GetSubtreeRoots_Handler(srv interface{}, stream grpc.Ser
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(CompactTxStreamerServer).GetSubtreeRoots(m, &compactTxStreamerGetSubtreeRootsServer{stream})
+	return srv.(CompactTxStreamerServer).GetSubtreeRoots(m, &grpc.GenericServerStream[GetSubtreeRootsArg, SubtreeRoot]{ServerStream: stream})
 }
 
-type CompactTxStreamer_GetSubtreeRootsServer interface {
-	Send(*SubtreeRoot) error
-	grpc.ServerStream
-}
-
-type compactTxStreamerGetSubtreeRootsServer struct {
-	grpc.ServerStream
-}
-
-func (x *compactTxStreamerGetSubtreeRootsServer) Send(m *SubtreeRoot) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetSubtreeRootsServer = grpc.ServerStreamingServer[SubtreeRoot]
 
 func _CompactTxStreamer_GetAddressUtxos_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(GetAddressUtxosArg)
@@ -978,21 +786,11 @@ func _CompactTxStreamer_GetAddressUtxosStream_Handler(srv interface{}, stream gr
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(CompactTxStreamerServer).GetAddressUtxosStream(m, &compactTxStreamerGetAddressUtxosStreamServer{stream})
+	return srv.(CompactTxStreamerServer).GetAddressUtxosStream(m, &grpc.GenericServerStream[GetAddressUtxosArg, GetAddressUtxosReply]{ServerStream: stream})
 }
 
-type CompactTxStreamer_GetAddressUtxosStreamServer interface {
-	Send(*GetAddressUtxosReply) error
-	grpc.ServerStream
-}
-
-type compactTxStreamerGetAddressUtxosStreamServer struct {
-	grpc.ServerStream
-}
-
-func (x *compactTxStreamerGetAddressUtxosStreamServer) Send(m *GetAddressUtxosReply) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CompactTxStreamer_GetAddressUtxosStreamServer = grpc.ServerStreamingServer[GetAddressUtxosReply]
 
 func _CompactTxStreamer_GetLightdInfo_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(Empty)
